Moved 'TimeParts' into the utility module
