+++ src/main/core/logger/shadow_logger.rs
@@ -1,6 +1,7 @@
 use crate::core::support::emulated_time::{self, EmulatedTime};
 use crate::core::worker::Worker;
 use crate::host::host::HostInfo;
+use crate::utility::time::TimeParts;
 use crossbeam::queue::ArrayQueue;
 use log::{Level, Log, Metadata, Record, SetLoggerError};
 use log_bindings as c_log;
@@ -26,57 +27,6 @@ const SYNC_FLUSH_QD_LINES_THRESHOLD: usize = 10 * ASYNC_FLUSH_QD_LINES_THRESHOLD
 const MIN_FLUSH_FREQUENCY: Duration = Duration::from_secs(10);
 
 static SHADOW_LOGGER: Lazy<ShadowLogger> = Lazy::new(|| ShadowLogger::new());
-
-/// Helper for formatting times.
-#[derive(Debug, Eq, PartialEq)]
-struct TimeParts {
-    hours: u32,
-    mins: u32,
-    secs: u64,
-    nanos: u64,
-}
-
-impl TimeParts {
-    fn from_nanos(total_nanos: u128) -> Self {
-        // Total number of integer seconds.
-        let whole_secs = u64::try_from(total_nanos / 1_000_000_000).unwrap();
-        // Total number of integer minutes.
-        let whole_mins = u32::try_from(whole_secs / 60).unwrap();
-        // Total number of integer hours, which is also the hours part.
-        let whole_hours = whole_mins / 60;
-
-        // Integer minutes, after whole hours are subtracted out.
-        let mins_part = whole_mins - whole_hours * 60;
-        // Integers secs, after integer minutes are subtracted out.
-        let secs_part = whole_secs - u64::from(whole_mins) * 60;
-        // Nanos, after integer secs are subtracted out.
-        let nanos_part =
-            u64::try_from(total_nanos - u128::from(whole_secs) * 1_000_000_000).unwrap();
-
-        Self {
-            hours: whole_hours,
-            mins: mins_part,
-            secs: secs_part,
-            nanos: nanos_part,
-        }
-    }
-}
-
-#[cfg(test)]
-#[test]
-fn test_time_parts() {
-    assert_eq!(
-        TimeParts::from_nanos(
-            (Duration::from_nanos(1) + Duration::from_secs(3600 + 60 + 1)).as_nanos()
-        ),
-        TimeParts {
-            hours: 1,
-            mins: 1,
-            secs: 1,
-            nanos: 1
-        }
-    );
-}
 
 /// Initialize the Shadow logger.
 pub fn init() -> Result<(), SetLoggerError> {
