+++ src/main/utility/mod.rs
@@ -9,6 +9,7 @@ pub mod pod;
 pub mod proc_maps;
 pub mod stream_len;
 pub mod syscall;
+pub mod time;
 
 /// A type that allows us to make a pointer Send + Sync since there is no way
 /// to add these traits to the pointer itself.
